test(native): add coverage for cache behavior
Move native cache_keep and cache_consent coverage into the cache
integration test suite, and add native backend coverage for max items
and age limit pruning tests.
